chore: revert textfield export, move textarea component
