+++ packages/components/src/index.ts
@@ -23,7 +23,6 @@ export * from './spinner'
 export * from './table'
 export * from './tag'
 export * from './tabs'
-export * from './textarea'
 export * from './textfield'
 export * from './toast'
 export * from './tooltip'
