--- packages/components/src/textarea/index.ts
@@ -1,1 +1,0 @@
-export {TextArea} from './TextArea';
