+++ packages/components/src/textfield/TextField.module.css
@@ -1,5 +1,5 @@
 @value tokens: "../theme/tokens.css";
-@value --text-disabled, --font-family from tokens;
+@value --focus, --field-disabled, --text-primary, --text-invalid, --border-invalid, --border-strong, --field-01, --field-hover-01, --font-family, --text-disabled from tokens;
 
 .textField {
   font-family: --font-family;
@@ -25,5 +25,52 @@
 
   &[data-disabled] {
     color: --text-disabled;
+  }
+}
+
+.textArea {
+  font-family: --font-family;
+  all: unset;
+  display: flex !important;
+  grid-area: Input;
+  border-bottom: 1px solid --border-strong;
+  align-items: center;
+  justify-content: center;
+  font-size: 1rem;
+  padding: 1rem 1rem 1.25rem;
+  background-color: --field-01;
+  box-sizing: border-box;
+  word-wrap: break-word;
+  width: 100%;
+  color: --text-primary;
+
+  @media (forced-colors: active) {
+    outline: 3px solid;
+    outline-offset: 2px;
+  }
+
+  &[data-hovered] {
+    background-color: --field-hover-01;
+  }
+
+  &[data-focused],
+  &[data-focus-visible] {
+    box-shadow: --focus;
+
+    @media (forced-colors: active) {
+      border: 1px solid;
+    }
+  }
+
+  &[data-invalid] {
+    border: none;
+    box-shadow: inset 0 0 0 3px --border-invalid;
+  }
+
+  &[data-disabled] {
+    cursor: not-allowed;
+    background-color: --field-disabled;
+    border: none;
+    opacity: 0.7;
   }
 }
