+++ packages/components/src/textfield/TextField.tsx
@@ -1,43 +1,15 @@
 'use client'
 
 import * as React from 'react'
-import {
-  TextField as AriaTextField,
-  TextFieldProps,
-  useContextProps,
-  TextFieldContext,
-} from 'react-aria-components'
-import styles from './TextField.module.css'
+import { TextFieldBase, type TextFieldBaseProps } from './TextFieldBase'
+import { Input } from '../input'
 import clsx from 'clsx'
 
-export type { TextFieldProps }
-
-export const TextField = React.forwardRef<HTMLDivElement, TextFieldProps>(
-  (props, ref) => {
-    ;[props, ref] = useContextProps(props, ref, TextFieldContext)
-
-    const [value, setValue] = React.useState(
-      props.defaultValue ?? props.value ?? '',
-    )
-
-    const handleChange: TextFieldProps['onChange'] = value => {
-      setValue(value)
-    }
-
-    const handleFocus: React.FocusEventHandler<HTMLInputElement> = event => {
-      setValue(event.target.value)
-    }
-
-    return (
-      <AriaTextField
-        {...props}
-        className={clsx(styles.textField, props.className)}
-        onChange={handleChange}
-        onFocus={handleFocus}
-        value={value}
-      >
-        {props.children}
-      </AriaTextField>
-    )
-  },
+export const TextField: React.FC<TextFieldBaseProps> = ({
+  className,
+  ...props
+}) => (
+  <TextFieldBase {...props}>
+    <Input className={clsx(className)} />
+  </TextFieldBase>
 )
