+++ packages/components/src/textfield/index.ts
@@ -1,1 +1,3 @@
+export * from './TextFieldBase'
 export * from './TextField'
+export * from './TextArea'
